Added power color text update to copying card function and removed clunkiness when dragging after enlarging a card
previously, if you clicked and dragged a card, the card would spawn after the drag function tried to destroy it. Now, we will check if isDragging is true after creating a card and will despawn it before it renders.

Also, we will now enlarge the base card after creating an enlarged copy so that the onPointerExit lines up with the enlarged card