Split loot deck into shop Loot and reward Loot
Made it so that the reward loot does not diminish, there will now be a static chance of loot drops and the loot will not run out.

Also modified Draft Card logic to assign to draft card layer when drawing a card from deck and then reassigning it back to the proper layer when purchased.

This prevents cards rewarded from loot from spawning as a draft card.