WIP smooth hand sorting
Currently, players are able to drag a card into hand and swap/sort as expected, but bugs arise when quickly exiting the hand, for example, a hand card getting despawned.

This seems to happen when the card exits the collider mid swap, so the swap sees no dragged card, and so it changes the position of the two cards, and then the previous slot is despawned.